requirements.txt
Flask==3.0.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
pytesseract==0.3.10
reportlab==4.0.4
Werkzeug==2.3.8